refactoring : fileLoad 컨트롤러 분리
