+++ src/main/java/webserver/RequestMappingManager.java
@@ -1,12 +1,10 @@
 package webserver;
 
 import controller.Controller;
-import http.FileExtension;
 import http.request.HttpRequest;
 import http.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import utils.FileIoUtils;
 
 import java.io.File;
 import java.util.Arrays;
@@ -52,41 +50,13 @@ public class RequestMappingManager {
 
         Controller controller = RequestMappingManager.getController(path);
         if (controller == null) {
-
-            FileExtension fileExtension = getExtension(path);
-            String filePath = fileExtension.getPhysicalPath().concat(path);
-            byte[] body = fileLoadFromPath(filePath);
-
-            if (body.length == 0) {
-                httpResponse.response400Header();
-                return;
-            }
-
-            httpResponse.addHeader("Content-Type", fileExtension.getMimeType() + ";charset=utf-8");
-            httpResponse.response200Header(body.length);
-            httpResponse.responseBody(body);
-            return;
+            controller = RequestMappingManager.getController("");
         }
 
         controller.service(httpRequest, httpResponse);
     }
 
-    public static byte[] fileLoadFromPath(String path) {
-        try {
-            if (path.endsWith("/")) return new byte[0];
-            return FileIoUtils.loadFileFromClasspath(path);
-        } catch (Exception ex) {
-            logger.error(ex.getMessage());
-        }
-        return new byte[0];
-    }
 
-    private static FileExtension getExtension(final String filePath) {
-        if (filePath.contains("\\.")) return FileExtension.NONE;
-        String[] split = filePath.split("\\.");
-        String extension = split[split.length - 1];
-        return FileExtension.of(extension);
-    }
 
     private static List<String> getControllerPathFile(final String packageName) throws Exception {
         File dir = new File(ROOT_ABSOLUTE_PATH + packageName.replace(".", "/"));
